Классы разделены по разным файлам. Теперь первыми могут ходить только белые шашки
