+++ ConcoleCheck/Rules.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConcoleCheck
+{
+    class Rules
+    {
+        public void BeginningOfGame()
+        {
+            Console.WriteLine("Как играть:\n O - белые шашки\n @  - черные шашки\n");
+        }
+                        
+    }
+}
